examples: modify the caculation for max and min latency
The caculation for max and min latency values are not accurate,
for example, to caculate the max latency, it uses the average max
value as max value in a short period of time, this results in
a lower max value than the real one. fix this by examining each
of the max value and find the overall max one. same changes applied
to caculate min value.

Signed-off-by: Zenghui Shi <zshi@redhat.com>