Xenalib: Conversion to run under Python2 and Python3
There is need for a python2 and python3 compatible XenaPythonLib,
since many distributions have been switching to default with Python3.

The main difference between Python2 and Python3 is the basic types
that exist to deal with texts and bytes. Python3 introduces more
distinguishing string types, but that means the "str" usage for
"bytes-on-the-wire" type programming will run into errors doing
comparisons.

Unfortunately, Python3 doesn't provide a convenient way to interpret
all strings as a particular type, so the conversion looks a bit ugly.

Some of this conversion was provided by the "2to3" utility.

Signed-off-by: Nandan Kulkarni <nakulkar@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>
Signed-off-by: Flavio Leitner <fbl@redhat.com>